+++ Source/Tests/Core/TestFailureHandlerChainTests.m
@@ -19,6 +19,22 @@
     [super tearDown];
 }
 
+- (void)testDefaultChain_ShouldPointToXCTestHandlerAsHeadOfChain
+{
+    HCTestFailureHandler *handler = [HCTestFailureHandlerChain chain];
+
+    STAssertEqualObjects(NSStringFromClass([handler class]), @"HCXCTestFailureHandler", nil);
+    STAssertNotNil(handler.successor, nil);
+}
+
+- (void)testDeprecatedFunction_ShouldPointToXCTestHandlerAsHeadOfChain
+{
+    HCTestFailureHandler *handler = HC_testFailureHandlerChain();
+
+    STAssertEqualObjects(NSStringFromClass([handler class]), @"HCXCTestFailureHandler", nil);
+    STAssertNotNil(handler.successor, nil);
+}
+
 - (void)testAddHandler_ShouldSetHeadOfChainToGivenHandler
 {
     HCTestFailureHandler *newHandler = [[HCTestFailureHandler alloc] init];